4.1 Using Templates
Configured a TEMPLATE_DIR variable with os.path.oin  and then supplied it in the list of template dirs for our project. when that was done we created tampltes/ and templates/rango to store index.html which was our first template with jinja. then we went through retunring the template viewwith render and found out about template context. 

I got confused about where to put the folder cause I wasn't paying attention. Shame on me.