8.1 Using Relative URLs in Templates
in this chapeter we went of the the syntax for using relative urls that rely on the framework to interpret vs actual URLs. the syntax is:

<a href={% url 'rango:about %}>about</a>
and 
<a href="{% url rango.views.about %}>About</a>

It then explains giving an app name space to your urls so that you can use the first syntax form above, to do this you add

app_name = <app_name> to your urls.py file for the app in you rproject and that enables you to use relative links like <a href="{% url 'rango:about' %}>About</a>

That's going to be my preferred way of doing things in the future.