8.2 + 8.3 with Exercises
This chapter taught us how to extend templates by creating a base.html template in the rango app and then extending it with our existing templates with 
{% extends 'rango.template.html' %}

inside the templates we de-emphasized html for jinija template language and practice implementing relative links, so <a href="{% url 'rango:about' %}>about</a> for example. from here all challenges are complete.