Chapter 8 Complete
We learned how to write custom template tags. To do this we create a templatetags module in our application with an init.py file and a file for our tag templates, in this case we called ours rango_template_tags.

inside this file, we  import template and use the template.library() class to add the template tags we've defined as functions to our applications list of  available template tags. 

The function explains how a template tag should work in conjuction with which template ir should use to render the tag. so in our template for the template tag we do a simple iterative loop that boldens the category we're currently viewing. 

I don't know that I will be using this feature of the framework but I think I should be able to recall it later.